+++ src/main/java/cs/csula/edu/cloudservice/entity/event/PositionEvent.java
@@ -18,9 +18,9 @@ public class PositionEvent {
   @GeneratedValue(strategy = GenerationType.AUTO)
   private UUID id;
 
-  private int x;
-  private int y;
-  private int z;
+  private float x;
+  private float y;
+  private float z;
 
   @ManyToOne
   @JoinColumn
@@ -40,27 +40,27 @@ public class PositionEvent {
     this.id = id;
   }
 
-  public int getX() {
+  public float getX() {
     return x;
   }
 
-  public void setX(int x) {
+  public void setX(float x) {
     this.x = x;
   }
 
-  public int getY() {
+  public float getY() {
     return y;
   }
 
-  public void setY(int y) {
+  public void setY(float y) {
     this.y = y;
   }
 
-  public int getZ() {
+  public float getZ() {
     return z;
   }
 
-  public void setZ(int z) {
+  public void setZ(float z) {
     this.z = z;
   }
